Extract class MyOpenLabDriver from DriverManager
Fixes java.lang.ClassNotFoundException
